refactor: Put vo package in domain package
