+++ src/main/kotlin/racingcar/domain/Car.kt
@@ -1,12 +1,10 @@
 package racingcar.domain
 
-import racingcar.vo.Name
-import racingcar.vo.Position
+import racingcar.domain.vo.Name
+import racingcar.domain.vo.Position
 
 class Car(name: String, position: Position = Position()) {
-
     val name = Name(name)
-
     var position = position
         private set
