+++ src/main/kotlin/racingcar/domain/RacingGame.kt
@@ -1,17 +1,16 @@
 package racingcar.domain
 
-import racingcar.vo.Name
-import racingcar.vo.Position
-import racingcar.vo.Round
+import racingcar.domain.vo.Name
+import racingcar.domain.vo.Position
+import racingcar.domain.vo.Round
 
 class RacingGame(private val cars: List<Car>, private var round: Round) {
-
-    constructor(names: List<String>, round: Int) : this(names.map { Car(it) }, Round(round))
-
     init {
         validateEmptyCars()
         validateNoRound()
     }
+
+    constructor(names: List<String>, round: Int) : this(names.map { Car(it) }, Round(round))
 
     private fun validateEmptyCars() {
         require(cars.size >= MIN_CAR_COUNT) { "at least 2 cars are needed to play" }
