+++ src/main/kotlin/racingcar/domain/vo/Position.kt
@@ -1,4 +1,4 @@
-package racingcar.vo
+package racingcar.domain.vo
 
 import racingcar.domain.MoveStrategy
