+++ src/test/kotlin/racingcar/domain/RacingGameTest.kt
@@ -6,8 +6,8 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.collections.shouldContainExactly
 import io.kotest.matchers.shouldBe
-import racingcar.vo.Name
-import racingcar.vo.Round
+import racingcar.domain.vo.Name
+import racingcar.domain.vo.Round
 
 internal class RacingGameTest : FunSpec({
